Added correct handling for enum's listed in an enum definition (CtEnum)
- before the toString() method was simply used which does not take
account of the compilation unit's context.
* Removed the compute imports statement from CtElement's toString()
method as it has no context and therefore it cannot assume the
corresponding import is present that would allow it to be shown in
the shortened version.
* Added test to catch some edge cases in the shortening of identifiers
that may produce ambiguous/invalid code and updated ones that relied on
the toString() method to calculate the shorthand names.